max_output_amplification_factor size_t->int64_t

+++ be/src/exec/pipeline/adaptive/collect_stats_context.h
@@ -76,7 +76,7 @@ public:
     void set_downstream_dop(size_t downstream_dop) { _downstream_dop = downstream_dop; }
     void incr_sinker() { ++_upstream_dop; }
 
-    const size_t max_output_amplification_factor() const { return _max_output_amplification_factor; }
+    const int64_t max_output_amplification_factor() const { return _max_output_amplification_factor; }
 
 private:
     using BufferChunkQueue = std::queue<ChunkPtr>;
@@ -102,7 +102,7 @@ private:
     size_t _downstream_dop = 0;
 
     const size_t _max_block_rows_per_driver_seq;
-    const size_t _max_output_amplification_factor;
+    const int64_t _max_output_amplification_factor;
 
     std::vector<BufferChunkQueue> _buffer_chunk_queue_per_driver_seq;
     std::vector<uint8_t> _is_finishing_per_driver_seq;
